make state float64 instead of float32

+++ HSP2/om.py
@@ -165,7 +165,7 @@ def state_om_model_run_prep(state, io_manager, siminfo):
     state['model_object_cache'] = model_object_cache
     state['model_exec_list'] = np.asarray(model_exec_list, dtype="i8") 
     if ModelObject.ops_data_type == 'ndarray':
-        state['state_ix'] = np.asarray(list(state['state_ix'].values()), dtype="float32")
+        state['state_ix'] = np.asarray(list(state['state_ix'].values()), dtype="float64")
     state['op_tokens'] = op_tokens 
     if len(op_tokens) > 0:
         state['state_step_om'] = 'enabled'
